Move ReaderModeButton into BrowserLocationView, private.
