+++ Client/Frontend/Browser/BrowserLocationView.swift
@@ -128,3 +128,37 @@ class BrowserLocationView : UIView, UIGestureRecognizerDelegate {
         }
     }
 }
+
+private class ReaderModeButton: UIButton {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setImage(UIImage(named: "reader.png"), forState: UIControlState.Normal)
+        setImage(UIImage(named: "reader_active.png"), forState: UIControlState.Selected)
+    }
+    
+    required init(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    var _readerModeState: ReaderModeState = ReaderModeState.Unavailable
+    
+    var readerModeState: ReaderModeState {
+        get {
+            return _readerModeState;
+        }
+        set (newReaderModeState) {
+            _readerModeState = newReaderModeState
+            switch _readerModeState {
+            case .Available:
+                self.enabled = true
+                self.selected = false
+            case .Unavailable:
+                self.enabled = false
+                self.selected = false
+            case .Active:
+                self.enabled = true
+                self.selected = true
+            }
+        }
+    }
+}
